+++ src/main/java/com/github/alexdlaird/ngrok/installer/NgrokV3CDNUrl.java
@@ -16,15 +16,15 @@ public enum NgrokV3CDNUrl implements NgrokCDNUrl {
     WINDOWS_i386("ngrok-v3-stable-windows-386.zip"),
     WINDOWS_x86_64("ngrok-v3-stable-windows-amd64.zip"),
     WINDOWS_x86_64_arm("ngrok-v3-stable-windows-arm64.zip"),
-    LINUX_i386("ngrok-v3-stable-linux-386.zip"),
-    LINUX_i386_arm("ngrok-v3-stable-linux-arm.zip"),
-    LINUX_x86_64("ngrok-v3-stable-linux-amd64.zip"),
-    LINUX_x86_64_arm("ngrok-v3-stable-linux-arm64.zip"),
-    FREEBSD_i386("ngrok-v3-stable-freebsd-386.zip"),
-    FREEBSD_i386_arm("ngrok-v3-stable-freebsd-arm.zip"),
-    FREEBSD_x86_64("ngrok-v3-stable-freebsd-amd64.zip");
+    LINUX_i386("ngrok-v3-stable-linux-386.tgz"),
+    LINUX_i386_arm("ngrok-v3-stable-linux-arm.tgz"),
+    LINUX_x86_64("ngrok-v3-stable-linux-amd64.tgz"),
+    LINUX_x86_64_arm("ngrok-v3-stable-linux-arm64.tgz"),
+    FREEBSD_i386("ngrok-v3-stable-freebsd-386.tgz"),
+    FREEBSD_i386_arm("ngrok-v3-stable-freebsd-arm.tgz"),
+    FREEBSD_x86_64("ngrok-v3-stable-freebsd-amd64.tgz");
 
-    private static final String CDN_URL_PREFIX = "https://bin.equinox.io/c/bNyj1mQVY4c/";
+    private static final String CDN_URL_PREFIX = "https://bin.ngrok.com/c/bNyj1mQVY4c/";
 
     private final String url;